+++ src/js/core/methods/index.js
@@ -28,6 +28,7 @@ import GetAddress from './GetAddress';
 import GetDeviceState from './GetDeviceState';
 import GetFeatures from './GetFeatures';
 import GetPublicKey from './GetPublicKey';
+import GetSettings from './GetSettings';
 import LiskGetAddress from './LiskGetAddress';
 import LiskGetPublicKey from './LiskGetPublicKey';
 import LiskSignMessage from './LiskSignMessage';
@@ -76,6 +77,7 @@ const classes: {[k: string]: any} = {
     'getDeviceState': GetDeviceState,
     'getFeatures': GetFeatures,
     'getPublicKey': GetPublicKey,
+    'getSettings': GetSettings,
     'liskGetAddress': LiskGetAddress,
     'liskGetPublicKey': LiskGetPublicKey,
     'liskSignMessage': LiskSignMessage,